+++ BitFaster.Caching.Benchmarks/Lfu/SketchIncrement.cs
@@ -1,56 +1,39 @@
-﻿
+﻿using System;
 using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Jobs;
 using BitFaster.Caching.Lfu;
+using MathNet.Numerics.LinearAlgebra.Factorization;
 
 namespace BitFaster.Caching.Benchmarks.Lfu
 {
     [SimpleJob(RuntimeMoniker.Net60)]
     [MemoryDiagnoser(displayGenColumns: false)]
     [HideColumns("Median", "RatioSD")]
-    public class Sketch
+    public class SketchIncrement
     {
         private static CmSketch<int, DisableAvx2> std = new CmSketch<int, DisableAvx2>(10, EqualityComparer<int>.Default);
         private static CmSketch<int, DetectAvx2> avx = new CmSketch<int, DetectAvx2>(10, EqualityComparer<int>.Default);
 
-        [GlobalSetup]
-        public void Setup()
-        {
-            for (int i = 0; i < 128; i++)
-            {
-                if (i % 3 == 0)
-                {
-                    std.Increment(i);
-                    avx.Increment(i);
-                }
-            }
-        }
-
         [Benchmark(Baseline = true)]
-        public int EstimateFrequency()
+        public void Inc()
         {
-            int count = 0;
             for (int i = 0; i < 128; i++)
             {
-                if (std.EstimateFrequency(i) > std.EstimateFrequency(i + 1))
-                    count++;
+                std.Increment(i);
             }
-
-            return count;
         }
 
         [Benchmark()]
-        public int EstimateFrequencyAvx()
+        public void IncAvx()
         {
-            int count = 0;
             for (int i = 0; i < 128; i++)
             {
-                if (avx.EstimateFrequency(i) > avx.EstimateFrequency(i + 1))
-                    count++;
+                avx.Increment(i);
             }
-
-            return count;
         }
     }
 }
