+++ BitFaster.Caching.UnitTests/Lfu/CMSketchAvx2Tests.cs
@@ -1,0 +1,7 @@
+﻿
+namespace BitFaster.Caching.UnitTests.Lfu
+{
+    public class CMSketchAvx2Tests : CmSketchTestBase<DetectAvx2>
+    {
+    }
+}
