+++ BitFaster.Caching.UnitTests/Lfu/CmSketchTests.cs
@@ -1,88 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using BitFaster.Caching.Lfu;
-using FluentAssertions;
-using Xunit;
-
+﻿
 namespace BitFaster.Caching.UnitTests.Lfu
 {
-    public class CmSketchTests
+    public class CmSketchTests : CmSketchTestBase<DisableAvx2>
     {
-        private CmSketch<int, DetectAvx2> sketch = new CmSketch<int, DetectAvx2>(512, EqualityComparer<int>.Default);
-
-        [Fact]
-        public void WhenCapacityIsZeroDefaultsSelected()
-        {
-            sketch = new CmSketch<int, DetectAvx2>(0, EqualityComparer<int>.Default);
-
-            sketch.ResetSampleSize.Should().Be(10);
-        }
-
-        [Fact]
-        public void WhenIncrementedOnceCountIsOne()
-        {
-            sketch.Increment(1);
-            sketch.EstimateFrequency(1).Should().Be(1);
-        }
-
-        [Fact]
-        public void WhenIncrementedMoreThanMaxCountIsMaximum()
-        {
-            for (int i = 0; i < 20; i++)
-            {
-                sketch.Increment(1);
-            }
-
-            sketch.EstimateFrequency(1).Should().Be(15);
-        }
-
-        [Fact]
-        public void WhenTwoItemsIncrementedCountIsIndependent()
-        {
-            sketch.Increment(1);
-            sketch.Increment(1);
-            sketch.Increment(2);
-
-            sketch.EstimateFrequency(1).Should().Be(2);
-            sketch.EstimateFrequency(2).Should().Be(1);
-        }
-
-        [Fact]
-        public void WhenSampleSizeExceededCountIsReset()
-        {
-            bool reset = false;
-
-            sketch = new CmSketch<int, DetectAvx2>(64, EqualityComparer<int>.Default);
-
-            for (int i = 1; i < 20 * 64; i++)
-            {
-                sketch.Increment(i);
-
-                if (sketch.Size != i)
-                {
-                    reset = true;
-                    break;
-                }
-            }
-
-            reset.Should().BeTrue();
-            sketch.Size.Should().BeLessThan(10 * 64);
-        }
-
-        [Fact]
-        public void WhenClearedCountIsReset()
-        {
-            sketch.Increment(1);
-            sketch.Increment(1);
-            sketch.Increment(2);
-
-            sketch.Clear();
-
-            sketch.EstimateFrequency(1).Should().Be(0);
-            sketch.EstimateFrequency(2).Should().Be(0);
-        }
     }
 }
